add func_activation_converter
