+++ python/paddle/tensorrt/impls/activation.py
@@ -16,6 +16,7 @@ import numpy as np
 import tensorrt as trt
 
 from paddle.tensorrt.converter_utils import (
+    activation_converter,
     get_trt_plugin,
 )
 from paddle.tensorrt.register import converter_registry
@@ -23,8 +24,12 @@ from paddle.tensorrt.register import converter_registry
 
 @converter_registry.register("pd_op.relu", trt_version="8.x")
 def relu_converter(network, paddle_op, inputs):
-    relu_layer = network.add_activation(inputs[0], trt.ActivationType.RELU)
-    return relu_layer.get_output(0)
+    return activation_converter(network, paddle_op.name(), inputs)
+
+
+@converter_registry.register("pd_op.tanh", trt_version="8.x")
+def tanh_converter(network, paddle_op, inputs):
+    return activation_converter(network, paddle_op.name(), inputs)
 
 
 @converter_registry.register("pd_op.softmax", trt_version="8.x")
