+++ test/tensorrt/test_converter_math.py
@@ -97,17 +97,5 @@ class TestAddTRTPattern(TensorRTBaseTest):
         self.check_trt_result()
 
 
-class TestTANHTRTPattern(TensorRTBaseTest):
-    def setUp(self):
-        self.python_api = paddle.tanh
-        self.api_args = {"x": np.random.randn(3).astype(np.float32)}
-        self.program_config = {"feed_list": ["x"]}
-        self.min_shape = {"x": [1]}
-        self.max_shape = {"x": [5]}
-
-    def test_trt_result(self):
-        self.check_trt_result()
-
-
 if __name__ == '__main__':
     unittest.main()
